move Camera_Init func to cam module
