+++ Modules/timers/timers.c
@@ -14,12 +14,15 @@
 #include "can/can.h"
 #include "timers.h"
 #include "lamp/lamp.h"
+#include "camera/camera.h"
 
-TIM_HandleTypeDef htim7;
 TIM_HandleTypeDef htim4;
+TIM_HandleTypeDef htim7;
+TIM_HandleTypeDef htim16;
+
 extern struct commands uartCommands;
 extern MessageTypeDef UART_MessageRecieved;
-int cam_bridge[3] = { 0, 0, 0 };
+extern int cam_bridge[3];
 
 extern uint8_t Counter_red;
 extern uint8_t Counter_blue;
@@ -27,26 +30,6 @@ extern uint8_t Counter_green;
 extern uint8_t maxCounterRed;
 extern uint8_t maxCounterBlue;
 extern uint8_t maxCounterGreen;
-
-void Camera_Init(void) {
-
-	GPIO_InitTypeDef GPIO_InitStruct = { 0 };
-
-	/* GPIO Ports Clock Enable */
-	__HAL_RCC_GPIOC_CLK_ENABLE();
-
-	/*Configure GPIO pin Output Level */
-	HAL_GPIO_WritePin(GPIOC, CAM_STER3_Pin | CAM_STER2_Pin | CAM_STER1_Pin,
-			GPIO_PIN_RESET);
-
-	/*Configure GPIO pins : PCPin PCPin PCPin */
-	GPIO_InitStruct.Pin = CAM_STER3_Pin | CAM_STER2_Pin | CAM_STER1_Pin;
-	GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-	GPIO_InitStruct.Pull = GPIO_NOPULL;
-	GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-	HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-}
 
 void TIM4_Init() {
 
@@ -135,17 +118,17 @@ void MX_TIM16_Init(void) {
 }
 
 void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim) {
-	if (htim->Instance == TIM7) {
-		if (uartCommands.numberOfCommands > 0) {
-			--uartCommands.numberOfCommands;
-		}
-		return;
-	}
-
 	if (htim->Instance == TIM4) {
 		HAL_GPIO_WritePin(CAM_STER1_GPIO_Port, CAM_STER1_Pin, 1);
 		HAL_GPIO_WritePin(CAM_STER2_GPIO_Port, CAM_STER2_Pin, 1);
 		HAL_GPIO_WritePin(CAM_STER3_GPIO_Port, CAM_STER3_Pin, 1);
+		return;
+	}
+
+	if (htim->Instance == TIM7) {
+		if (uartCommands.numberOfCommands > 0) {
+			--uartCommands.numberOfCommands;
+		}
 		return;
 	}
 
@@ -178,7 +161,6 @@ void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim) {
 
 void HAL_TIM_OC_DelayElapsedCallback(TIM_HandleTypeDef *htim) {
 	if (htim == &htim4) {
-
 		int cam_bridge_expected;
 
 		switch (HAL_TIM_GetActiveChannel(&htim4)) {
