+++ Modules/timers/timers.h
@@ -16,12 +16,8 @@
 #define CAM_STER1_Pin GPIO_PIN_2
 #define CAM_STER1_GPIO_Port GPIOC
 
-//int cam_bridge[3];
-
-void TIM7_Init(void);
-
-void Camera_Init(void);
-
 void TIM4_Init(void);
+void TIM7_Init(void);
+void MX_TIM16_Init(void);
 
 #endif /* MODULES_TIMERS_TIMERS_H_ */
